refactor: 초기화에서 게임기록 관리 로직 분리
